+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1.1.0
 
 - Tap-based calculation of tempo through hitting ENTER key
+- Mute metronome
+- Stop/Start metronome
 
 ## 1.0.0